% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download}
\alias{download}
\title{General function for downloading data}
\usage{
download(type, ...)
}
\arguments{
\item{type}{a character value indicating the type of data eg. \code{"ctd"}}
}
\description{
bfbwyfbfbwhfyeerwhbdbgdfgf
}
\author{
Annie Howard, Jaimie Harbin
}
